[~] AppSettings: Refactor variable types for consistency [~] GUIMenuBar: Rename Shadowed Variables [~] LogicElementBase: Use `size_t` instead of `int`

+++ include/appSettings.h
@@ -24,10 +24,10 @@ namespace AppSettings
         bool showDemoWindow = false;
         int SPACING_SIZE = 25;
         int MOUSE_SELECTION_OFFSET = SPACING_SIZE / 2;
-        int GRID_POINT_SIZE = 3;
-        int SLICE_SIZE = 40;
-        float IN_OUT_RECT_WIDTH = static_cast<float>(GRID_POINT_SIZE * 5);
-        float IN_OUT_INTERACTION = static_cast<float>(IN_OUT_RECT_WIDTH * 2);
+        float GRID_POINT_SIZE = 3;
+        float SLICE_SIZE = 40;
+        float IN_OUT_RECT_WIDTH = GRID_POINT_SIZE * 5;
+        float IN_OUT_INTERACTION = IN_OUT_RECT_WIDTH * 2;
     };
 
     extern AppSettings::Settings appSettings;
